Fix some things, move to new file src/authentication_results_handling.rs
