+++ src/lib.rs
@@ -93,6 +93,7 @@ mod update_helper;
 pub mod webxdc;
 #[macro_use]
 mod dehtml;
+mod authentication_results_handling;
 mod color;
 pub mod html;
 pub mod plaintext;
