+++ src/sql/migrations.rs
@@ -599,10 +599,7 @@ CREATE INDEX smtp_messageid ON imap(rfc724_mid);
     if dbversion < 92 {
         info!(context, "[migration] v92");
         sql.execute_migration(
-            // TODO Is this really the database scheme we want?
-            // Would be possible to save the timestamp here until when it was correct (change it as soon as it becomes incorrect).
-            // Then if this is old enough, accept a deviating key again
-            "ALTER TABLE acpeerstates ADD COLUMN dkim_status INTEGER DEFAULT 0;",
+            "CREATE TABLE sending_domains(domain TEXT PRIMARY KEY, correct_dkim INTEGER DEFAULT 0);",
             92,
         )
         .await?;
